refactor: type error 해결
